Fixes:    Compilation on 64-bit architectures Details:  Hope it's the correct fix :)
[git-p4: depot-paths = "//depot/qt/main/": change = 234592]

+++ util/qdoc3/codemarker.cpp
@@ -103,8 +103,8 @@ const Node *CodeMarker::nodeForString( const QString& string )
 
 QString CodeMarker::stringForNode( const Node *node )
 {
-    if (sizeof(const Node *) == sizeof(uint)) {
-        return QString::number(reinterpret_cast<uint>(node));
+    if (sizeof(const Node *) == sizeof(ulong)) {
+        return QString::number(reinterpret_cast<ulong>(node));
     } else {
         return QString::number(reinterpret_cast<qulonglong>(node));
     }
